+++ tests/conftest.py
@@ -6,7 +6,12 @@ import responses
 from responses import BaseResponse, RequestsMock, matchers
 from sqlite_utils import Database
 
-from reddit_user_to_sqlite.reddit_api import USER_AGENT, Post, SuccessResponse
+from reddit_user_to_sqlite.reddit_api import (
+    USER_AGENT,
+    ErrorHeaders,
+    PagedResponse,
+    Post,
+)
 from reddit_user_to_sqlite.sqlite_helpers import CommentRow, PostRow, UserRow
 
 
@@ -26,7 +31,7 @@ def tmp_db(tmp_db_path):
     return Database(tmp_db_path)
 
 
-def _wrap_response(*children) -> SuccessResponse:
+def _wrap_response(*children) -> PagedResponse:
     return {
         "kind": "Listing",
         "data": {
@@ -274,7 +279,7 @@ def stored_removed_comment_placeholder_user() -> CommentRow:
 
 
 @pytest.fixture
-def comment_response(comment) -> SuccessResponse:
+def comment_response(comment) -> PagedResponse:
     """
     The full response from Reddit with a comment child
     """
@@ -652,6 +657,7 @@ class MockPagedFunc(Protocol):
         resource: Literal["comments", "submitted"],
         json: Any,
         params: Optional[dict[str, Union[str, int]]] = None,
+        headers: Optional[dict[str, str]] = None,
     ) -> BaseResponse:
         ...
 
@@ -666,6 +672,7 @@ def mock_paged_request(mock: RequestsMock) -> MockPagedFunc:
         resource: Literal["comments", "submitted"],
         json: Any,
         params: Optional[dict[str, Union[str, int]]] = None,
+        headers: Optional[dict[str, str]] = None,
     ):
         params = {"limit": 100, "raw_json": 1, **(params or {})}
 
@@ -676,13 +683,16 @@ def mock_paged_request(mock: RequestsMock) -> MockPagedFunc:
                 matchers.header_matcher({"user-agent": USER_AGENT}),
             ],
             json=json,
+            headers=headers,
         )
 
     return _mock_request
 
 
 class MockInfoFunc(Protocol):
-    def __call__(self, ids: str, json: Any, limit=100) -> BaseResponse:
+    def __call__(
+        self, ids: str, json: Any, headers: Optional[dict[str, str]] = None, limit=100
+    ) -> BaseResponse:
         ...
 
 
@@ -699,6 +709,7 @@ def mock_info_request(mock: RequestsMock) -> MockInfoFunc:
     def _mock_request(
         ids: str,
         json: Any,
+        headers: Optional[dict[str, str]] = None,
         limit=100,
     ):
         params = {"limit": limit, "raw_json": 1, "id": ids}
@@ -710,6 +721,7 @@ def mock_info_request(mock: RequestsMock) -> MockInfoFunc:
                 matchers.header_matcher({"user-agent": USER_AGENT}),
             ],
             json=json,
+            headers=headers,
         )
 
     return _mock_request
@@ -807,6 +819,15 @@ def user_response():
             "accept_followers": True,
             "has_subscribed": False,
         },
+    }
+
+
+@pytest.fixture
+def rate_limit_headers() -> ErrorHeaders:
+    return {
+        "x-ratelimit-used": "4",
+        "x-ratelimit-remaining": "6",
+        "x-ratelimit-reset": "20",
     }