+++ tests/test_reddit_api.py
@@ -1,10 +1,11 @@
-from unittest.mock import patch
+from unittest.mock import MagicMock, patch
 
 import pytest
 
 from reddit_user_to_sqlite.reddit_api import (
-    SuccessResponse,
-    _raise_reddit_error,
+    PagedResponse,
+    RedditRateLimitException,
+    _unwrap_response_and_raise,
     add_missing_user_fragment,
     get_user_id,
     load_comments_for_user,
@@ -20,6 +21,27 @@ def test_load_comments(mock_paged_request: MockPagedFunc, comment_response, comm
     assert load_comments_for_user("xavdid") == [comment]
 
     assert response.call_count == 1
+
+
+@patch("reddit_user_to_sqlite.reddit_api.PAGE_SIZE", new=1)
+def test_load_comments_rate_limited(
+    mock_paged_request: MockPagedFunc, comment_response, comment, rate_limit_headers
+):
+    good_response = mock_paged_request(
+        resource="comments", params={"limit": 1}, json=comment_response
+    )
+    bad_response = mock_paged_request(
+        resource="comments",
+        params={"limit": 1},
+        json={"error": 429},
+        headers=rate_limit_headers,
+    )
+
+    # despite getting an error, we still got the first comment
+    assert load_comments_for_user("xavdid") == [comment]
+
+    assert good_response.call_count == 1
+    assert bad_response.call_count == 1
 
 
 def test_load_posts(mock_paged_request: MockPagedFunc, self_post_response, self_post):
@@ -42,7 +64,7 @@ def test_loads_10_pages(mock_paged_request: MockPagedFunc, comment_response, com
 
 @patch("reddit_user_to_sqlite.reddit_api.PAGE_SIZE", new=1)
 def test_loads_multiple_pages(
-    mock_paged_request: MockPagedFunc, comment_response: SuccessResponse, comment
+    mock_paged_request: MockPagedFunc, comment_response: PagedResponse, comment
 ):
     comment_response["data"]["after"] = "abc"
     first_request = mock_paged_request(
@@ -96,19 +118,53 @@ def test_load_info_pages(mock_info_request: MockInfoFunc, comment_response, comm
     assert load_info(["a", "b", "c", "d", "e"]) == [comment] * 3
 
 
+@patch("reddit_user_to_sqlite.reddit_api.PAGE_SIZE", new=2)
+def test_load_info_pages_with_rate_limit(
+    mock_info_request: MockInfoFunc, comment_response, comment, rate_limit_headers
+):
+    mock_info_request("a,b", json=comment_response, limit=2)
+    mock_info_request("c,d", json=comment_response, limit=2)
+    mock_info_request("e", json={"error": 429}, limit=2, headers=rate_limit_headers)
+
+    # call for e fails, but we still got the first ones
+    assert load_info(["a", "b", "c", "d", "e"]) == [comment] * 2
+
+
 def test_load_info_empty(mock_info_request: MockInfoFunc, empty_response):
     mock_info_request("a,b,c,d,e,f,g,h", json=empty_response)
 
     assert load_info(["a", "b", "c", "d", "e", "f", "g", "h"]) == []
 
 
-def test_raise_reddit_error():
-    assert _raise_reddit_error({}) == None  # no raise, no return
+def test_unwrap_and_raise_passes_good_responses_through():
+    response = {"neat": True}
+    assert _unwrap_response_and_raise(MagicMock(json=lambda: response)) == response
 
+
+def test_unwrap_and_raise_raises_unknown_errors():
     with pytest.raises(ValueError) as err:
-        _raise_reddit_error({"error": 123, "message": "cool"})
-
+        _unwrap_response_and_raise(
+            MagicMock(json=lambda: {"error": 123, "message": "cool"})
+        )
     assert str(err.value) == "Received API error from Reddit (code 123): cool"
+
+
+def test_unwrap_and_raise_raises_rate_limit_errors(rate_limit_headers):
+    with pytest.raises(RedditRateLimitException) as err:
+        _unwrap_response_and_raise(
+            MagicMock(
+                json=lambda: {"error": 429, "message": "cool"},
+                headers=rate_limit_headers,
+            )
+        )
+
+    e = err.value
+
+    assert e.used == 4
+    assert e.remaining == 6
+    assert e.window_total == 10
+    assert e.reset_after_seconds == 20
+    assert e.stats == "Used 4/10 requests (resets in 20 seconds)"
 
 
 def test_get_user_id(mock_user_request: MockUserFunc, user_response):